+++ eslint-plugin/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   rules: {
     "no-errors": require("./rules/no-errors"),
+    "no-meta-mismatch": require("./rules/no-meta-mismatch"),
     "no-unused-meta": require("./rules/no-unused-meta")
   }
 };